+++ src/lib.rs
@@ -33,11 +33,7 @@
 #![cfg_attr(feature = "cargo-clippy", warn(clippy::pedantic))]
 #![cfg_attr(
     feature = "cargo-clippy",
-    allow(
-        clippy::should_implement_trait,
-        clippy::missing_errors_doc,
-        clippy::missing_panics_doc,
-    )
+    allow(clippy::should_implement_trait, clippy::missing_errors_doc,)
 )]
 
 extern crate hashlink;